Add authorizeuser to action helper
